Add delayed sending Action ExecuteSession to self

+++ contracts/galactic-express/contract/state/src/state.rs
@@ -6,7 +6,7 @@ use launch_io::*;
 pub struct ParticipantInfo {
     address: ActorId,
     name: String,
-    balance: u32,
+    balance: u128,
 }
 
 #[metawasm]
